Removed duplicate files and renamed Lavesh commit to make it easier to find it.
