--- POC1/templates/main.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>OpenASSA POC</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.0/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <script src="https://www.w3schools.com/lib/w3codecolor.js"></script>
-    <style>
-        tbody {
-        display: block;
-        overflow: auto;
-        max-height: 400px;
-        width: 90%;
-        }
-
-        thead tr {
-        display: table;
-        width: 90%;
-        table-layout: fixed;
-        }
-
-        thead, tbody tr {
-        display: table;
-        width: 90%;
-        table-layout: fixed;
-        }
-
-        thead {
-        width: calc( 90% - 1em )
-        }
-    </style>
-</head>
-<body>
-
-<div class="container">
-    <h2>Policy Liability</h2>
-
-    <ul class="nav nav-tabs">
-        <li class="active"><a data-toggle="tab" href="#menu1">Policy Data</a></li>
-        <li><a data-toggle="tab" href="#menu2">Mortality Table</a></li>
-        <li><a data-toggle="tab" href="#menu3">Lapse Table</a></li>
-        <li><a data-toggle="tab" href="#menu4">Yield Curve</a></li>
-        <li><a data-toggle="tab" href="#menu5">Rules</a></li>
-        <li><a data-toggle="tab" href="#menu6">Results</a></li>
-    </ul>
-
-    <div class="tab-content">
-        <div id="menu1" class="tab-pane fade in active">{{ tables['policies']|safe }}</div>
-        <div id="menu2" class="tab-pane fade">{{ tables['mortality']|safe }}</div>
-        <div id="menu3" class="tab-pane fade">{{ tables['lapse']|safe }}</div>
-        <div id="menu4" class="tab-pane fade">{{ tables['yieldc']|safe }}</div>
-        <div id="menu5" class="tab-pane fade">
-            <h4>Rules</h4>
-            {{ tables['rules']|safe }}
-            <h4>Rules - Key Errors</h4>
-            {{ tables['rules_keyerror']|safe }}
-            <h4>Rules - Init</h4>
-            {{ tables['rules_init']|safe }}
-            <h4>Rules - Init - Key Errors</h4>
-            {{ tables['rules_init_keyerror']|safe }}
-        </div>
-        <div id="menu6" class="tab-pane fade">{{ tables['result']|safe }}</div>
-    </div>
-</div>
-
-<script>
-    w3CodeColor();
-</script>
-
-</body>
-</html>
